New PlannerHandler class.
Starting, stopping, destroying the planner is now handled by the PlannerHandler class. Additionally, the opengl drawing is simplified. To render a scene from the worker thread (c++ planners, move3d code), simply call the draw_opengl() function. The call is executed in the Qt thread, but it is blocking, so it is safe to call it directly.