Improvment of the planning API and work on new MLP algorithms
It is now possible to copy a graph (for display in a multi thread context).
Add a new class for HRI navigation.
The replanning algorithm and simulators are set in two distinct classes.
Add a new function in confGenerator for computing an IK from an Eigen WSPoint.
New global variables.
Use the confPtr_t macro in most places.